Changes Added start of HelpCommands Modified csProj a bit (fixed includes) Added start of Help/*.md files Renamed GuildPrefix.cs to reflect class rename Turned off default help Fixed GuildOption not returning properly (should've caught that earlier, mybad) # Help files
Gets read in full, and put in an embed.
Be sure to not make it too large, or discord can't send it.
If the file starts with `http`, it will extract the entire first line, and set that as the image Url

So, we can do header-images in our md files.

`{action}` or {character}` gets replaced by their emoji.
Can do the same for {item} once I've got those changes 😉